Use service to deal with surfing ads instead of doing so in activity.
