+++ app/src/main/java/com/vincent/mfssurfing/OnSurfFinishedListener.kt
@@ -1,0 +1,7 @@
+package com.vincent.mfssurfing
+
+interface OnSurfFinishedListener {
+
+    fun finished()
+
+}
